feat. 리뷰반영: 도메인 로직 메인으로 이동
