+++ src/test/kotlin/study/DslTest.kt
@@ -91,53 +91,4 @@ class DslTest {
     private fun introduce(block: PersonBuilder.() -> Unit): Person {
         return PersonBuilder().apply(block).build()
     }
-
-    data class Person(
-        val name: String,
-        val company: String?,
-        val softSkills: List<String>?,
-        val hardSkills: List<String>?,
-        val languageSkills: List<Pair<String, Int>>?
-    )
-
-    //class Person { //불변성이 보장안됨
-    class PersonBuilder {
-        private lateinit var name: String
-        private var company: String? = null
-        private var softSkills: MutableList<String> = mutableListOf()
-        private var hardSkills: MutableList<String> = mutableListOf()
-        private var languageSkills: MutableList<Pair<String, Int>> = mutableListOf()
-
-        fun name(name: String) {
-            this.name = name
-        }
-
-        fun company(value: String) {
-            company = value
-        }
-
-        fun skills(block: PersonBuilder.() -> Unit) {
-            block()
-        }
-
-        fun soft(value: String) {
-            softSkills.add(value)
-        }
-
-        fun hard(value: String) {
-            hardSkills.add(value)
-        }
-
-        fun languages(block: PersonBuilder.() -> Unit) {
-            block()
-        }
-
-        infix fun String.level(level: Int) {
-            languageSkills.add(Pair(this, level))
-        }
-
-        fun build(): Person {
-            return Person(name, company, softSkills, hardSkills, languageSkills)
-        }
-    }
 }
